Skip c tests if CHPL_LAUMCHER!=none
c tests have no launcher, so they will always run locally or on a login node,
which doesn't really make sense.

This caused issues with performance testing on xc

Previously we were testing c tests for things like gasnet with the amudpt
launcher. The c test wasn't actually "launched" it was just run without a
launcher on the local machine. If there's any use to that, this check can be
updated to only skip c tests if launcher!=none for performance tests.