Enable tests with stdin for slurm
This allows us to test performance test like some of the shootouts during out
single node xc performance testing.

Apparently stdin redirection DOES work with slurm. I guess I thought that since
launchertimeout was required stdinredirect wasn't supported as well since I was
used to throwing those flags at the same time?

This updates start_test so it no longer throws nostdinredirect automatically
for slurm. It also updates chpl_launchcmd to copy stdin to a file, and set an
env var for slurm so it knowns where to look for stdin.

Interactive slurm (not using sbatch) will work with just '<' for redirection.
When sbatch is being used (like with chpl_launchcmd) stdin has to be specified.
In this case we specify the file to use for stdin redirection.

We could just tell slurm to use the actual .stdin in sub_test instead of giving
chpl_launchcmd the stdin, and then having it write the file out again just so
slurm can use it, but I think that logic would be more complicated in sub_test,
and it's not needed unless batch mode is being used, in which case regular
start_test/sub_test don't work anyways.

Tested w/ and w/o chpl_launchcmd for the hellos and a directory that contains a
bunch of .stdins. Also tested release/examples/ w/o chpl_launchcmd