Add stdout "buffering" to slurm
For single locale performance testing we have a couple tests that generate
massive output files (fasta, a writeln test, and a few others.)

For batch mode, slurm buffers stdout on a per line basis and syncs the stdout
buffer to a file on the login node every so often. This constant syncing really
slows down these programs. For instance fasta generates a ~250 MB file and
previously this test took ~100 seconds to run. By sending output to /tmp and
then copying it, the tests only takes ~4 seconds (roughly the same time it
takes on chap04!)

This buffering works by setting slurms --output to /tmp/<jobid>.out and then
after the job is completed moving the /tmp file to the output file that a user
specified or slurm auto-generated. In my (limited) testing, this also works for
multi-locale tests since all output is piped through a single node.

There's a pretty fat comment in in the slurm-srun launcher that goes into some
more detail.

I'm not sure if there's a similar issue for tests with a large stdin, but if
there is we can use slurms sbcast to take care of it.