Remove --verbose and --hostlist command line options from chpl_launchcmd
These options can still be set with env vars. Technically they can still be
specified on the command line using the env var name as the option. I did this
because it's easier, and it leaves the info about setting the env vars in the
--help output.

We're moving away from --verbose and --hostlist because some chapel programs
use these and chpl_launchcmd was intercepting these args so the chapel program
would never see them.

This was noticed with xc single node perf testing where some tests throw
--verbose=true. This causes chpl_launchcmd to gripe because it's verbose was
specified with --verbose (=true/false) and it didn't know how to parse the arg.