Save the sha for perf and compiler perf runs
We want to be able to map a specific date back to the sha easily. For now we're
just putting the long sha in a special .dat file. The long term plan is to
have gengraphs do something special with this dat file, and make the sha for
each date available to dygraphs. Dygraphs will then display the sha for the
day either in the legend above the date, or if you hover over a point.

This stores the long sha, since the short sha could become ambiguous over time.
My current thinking is that the .dat file will always store just the long sha,
and then gengraphs can run "git rev-parse --short <long sha>" every night. This
way we can display the short sha on the graphs, without having to worry about
them becoming ambiguous.