Finish effort to move docs pass to just after checkParsed.
The docs pass was after flattenClasses due to wanting inheritance to have
resolved to the parent types.  This functionality lived in scopeResolve, and
flattenClasses made things somewhat easier than just dealing with scopeResolve's
resultant AST.  However, since the inheritance portion of scopeResolve was
fairly independent of its other functionality, I decided that I could reuse
that portion in the docs pass without the other changes that occurred to the
AST by making these functions accessible outside of scopeResolve and calling
them during the docs pass.  This had two negative outcomes, which we find
acceptable: first, inheritance from the object class was ignored - the object class doesn't provide any functionality, so this is no loss; second,
calling these functions twice (once in the docs pass and once during
scopeResolve) caused the compiler to fail during the second call - this merely
prevents generating an executable for the file you have documented, which does
not seem like a very necessary use case.  In the latter outcome, one could
investigate the impact on scopeResolve of calling the function earlier than
expected and with that information make the function more robust to this usage,
but we didn't feel that was necessary at this point in time.

This and the previous (few) commits have had some impact on the chpldoc tests.
I will make another commit to update them accordingly.