Temporarily force slurm launcher to use /tmp for stdout buffer
I'm temporarily disabling the logic to figure out what tmpdir to use and
forcing it to just use '/tmp' to see if it resolves single local xc perf
testing issues. We set TMPDIR in one of our common scripts. We then make that
dir on the login node, but not on the compute nodes, so the program can't
actually put it's output anywhere. If this resolves the issues, I'll update
the launcher to do something smarter like create the temp dir before running
and remove it afterwards.