Attempt to use AstToText in chpldoc output.
Starts by switching the name generation and formal output for functions and
methods to purely relying on AstToText's appendNamesAndFormals.  While this
solves the queried types problem and the default values problem for formals, it
has the unfortunate side effect of breaking nine chpldoc tests.  These errors
can be divided into two categories: method name related, and argument type
output related.

The method name related errors are because AstToText always prefixes the method
name with the type name and a period, whether the method is a primary or a
secondary method - chpldoc intends to only prefix the method name in the case of
secondary methods, since output under the class name will make the type name
associated with it apparent.  This class of errors can easily be solved by
only using AstToText's appendFormals method instead of appendNameAndFormals if
it is believed that AstToText should always append that prefix.

The argument type output related errors vary from obviously wrong to perhaps a
matter of preference.  The two obviously wrong cases are ranges and domains: a
range used within an array specification turns from `[1..4]` to
`[1chpl_build_bounded_range4]`, while a domain turns from `domain(1)` to
`AppendExpr.Call02`.  In the final case, I expected homogenuous tuples to output
as `3*(string)` and `3*(4*(complex))`, but they were turned into `3*string` and
`3*4*complex` respectively.