Have the docs pass use AstToText to generate function names and arguments.
This removes the need for the functions printArgs and printIntents.  A couple of
bug fixes were needed in AstToText as a result of extended usage: the output for
ranges within array arguments, domains, and homogenous tuples were not ideal or
correct in some cases.  Additionally, I wanted the name output for methods to
distinguish between primary and secondary methods (methods that are defined
within the class definition, and those defined outside of it).  These changes
are now present in AstToText.

This fixes three bugs: arguments that query portions of their definition now
output correctly, default values for arguments are now printed, and arguments
with a default value but no defined type no longer output "_unknown" as their
type.