Attempt to make C-level assert() naming conflicts go away
For some reason, recently, and only sometimes, we've seen naming
conflicts between (as I understand it) chpl-generated assert()
routines and the C routine of the same name.

The assert() routine was in an internal module, so does not
(currently) get our default renaming.

This commit moves it to modules/standard in order to get the
munging and make it available to user codes (which already
make use of it apparently).  While this is arguably a lame
way to fix the issue, it seems likely that we want assert()
to be available to end-users and, to that end, the standard
modules seem like a better place for it.

Open question, though:  assert() seems to be on the reservedSymbolNames
list, so why would it not have already gotten renamed.