Make it more obvious when somebody breaks compiler performance testing
The test 'passCheck' was meant to alert us if compiler performance testing is
broken. The test served as a sentinel to let you know compiler performance .dat
and .graph files have to be updated. Just fixing the test didn't fix compiler
performance testing.

There was a README next to the test explaining this, but it was really easy to
miss the README especially since the "fix" for the test seems so obvious.

This removes the README and puts a short note in compilerPerformance.perfkeys,
which is the file you have to change for the test to pass. It would be nice to
make it read only as well as an extra indicator but git doesn't store file
permissions.

The prediff had to be updated to allow comments in the .perfkey file and I
cleaned up the prediff logic while I was there.t  I also added a much larger
comment in the prediff to better explain compiler performance testing for those
interested with pointers from the .perfkey file.