Add a (commented out) flag to skip the documentation of extern functions.
Most of this functionality was already present, but we had not created a flag
for it yet.  However, since we believe that any externs visible at the module
level in our standard modules should be in a smaller scope, we decided to leave
the default as is and not make the flag visible, thereby forcing the documenter
to make this change instead of giving them an easy way out.  We should discuss
further whether we would like this flag present (but since it was a Five Minute
Fix (TM), there isn't much harm in placing it in the repo so long as the work
has been done.)