Rename 'by' in the parser in order to avoid naming conflicts
In one of my working patches, I ran into an issue because a
test named 'by.chpl' which introduces a module named 'by'
conflicted with references to the 'by' operator in the
internal module I was working on.  While this surprised me
at first ("by" is a reserved keyword?!) it makes sense given
that we convert instances of the 'by' operator, including
user-defined overloads of it, into functions named 'by'.

This pull request renames 'by' to 'chpl_by' in the parser
to protect against such issues (since users should not name
modules chpl_<anything>).  It turned out there was also a
helper function in ChapelRange.chpl named 'chpl_by', so then
I had to rename that to 'chpl_by_help' (which seems like an
appropriate name).

I think this is a step forward, albeit a somewhat silly one.