Replace uses of sleep with sync variables
As has been pointed out in several different tests, using sleeps
to try and order tasks is an unsafe practice.  This test was
failing on my Mac using array views which seemed surprising
given that it doesn't even use arrays.  Turns out that it's
just a race condition.  Replacing the sleeps with sync variables
to avoid the race.

The comment states that this practice was done for debugging
which is all well and good when debugging, but once tests
are being committed for long-term testing, let's try and get
the potential for races out of them please.