Massage and comment iterators, part 1.
+++ b/compiler/AST/iterator.cpp
In isSingleLoopIterator(FnSymbol* fn, Vec<BaseAST*>& asts) {
noted that the multi-case conditional isForLoop() || isCForLoop || isWhileDoLoop appearing
in two places could be replaced by a dynamic_cast<>() call.  I used dynamic_cast instead
of the isLoopStmt macro for two reasons: 1) I am confident that it does what I want; and
2) it is almost certainly faster.
I think the resulting code is clearer, because it first select the general case of
interest and then weeds out the exceptions.

Added header comments to buildZip1.  Noted that dumping all of the local variables into
Zip1-4 is sloppy and suggested a way this could be done more surgically, factoring out the
common code.
For the loop test condition in a WhileLoop, we used to generate
  tmp <- _cond_test(condExpr)
  if (tmp)
    _this.more <- 1
  else
    _this.more <- 0
I leveraged the bool -> int coercion in the backend to write the simpler expression
  _this.more <- tmp
This needs to be tested with the LLVM backend.
Performed the same simplification in buildZip3.

In rebuildIterator(IteratorInfo* ii, enabled the noOtherCalls assertion with fVerify is true.

+++ b/compiler/resolution/functionResolution.cpp
In protoIteratorMethod(IteratorInfo* ii, const char* name, Type* retType) {
added comments to clarify how the defaultInitializer field is used for illicit purposes by
the iterator record and class definitions.
Noted that the addition of the FLAG_RESOLVED flag could be factored into
protoIteratorMethod().

+++ b/compiler/resolution/lowerIterators.cpp
Added a header comment to expandForLoop explaining what expanded single-yield iterators
look like.
Added more descriptions of how expandForLoop works.
In inlineIterators(), inverted the logic for identifying live blocks to instead weeding
them out.  The existing test did not have an associated comment.