Massage and comment iterators, part 2.
Added header comments to buildAdvance().
Removed the seemingly unnecessary copy of cforLoop->testBlockGet() in buildHasMore(),
buildInit() and buildIncr() as noted by mnoakes.
In insertLocalsForRefs(Vec<Symbol*>& syms, FnSymbol* fn, reworked the logic slightly, to
expose that whether the RHS of a move is a primitive or a function call is a subcase of
its being a call.  This permitted the insertion of a new INT_FATAL with more-useful error
messages in each (what is wrong with the expected structure of the AST at that point in
the execution).
Computation of se and call within their respective "if" clauses places them as close as
possible to their uses, and minimizes their respective scopes -- both elements of good
coding practice.