Make cwd a method on a locale
Updated the tests of cwd to either use here.cwd or a specified locale, as
appropriate.

Update cwd to be a method on a locale.  This involved wrapping the call to
obtain the current working directory (and to convert it from a c_string into a
Chapel string) in an on statement for the specified locale, so that calls such
as Locale[1].cwd would work as expected.