Update setchplenv.* tests to work when CHPL_HOME is a symlink.
Use os.stat() when comparing expected vs. actual directories. This will work
even if CHPL_HOME is a symlink, since the stat values are the same for symlinks
and their targets.

Note that this happens because subprocess.Popen() sets PWD in the environment
to the real path instead the symlink path.