Speed up multi-config graphs
We use multi-config graphs for comparing multiple configurations of a test on
the same graph such as comparing ugni vs. gn-mpi vs gn-aries

You can toggle which configs you want to view with check boxes, and then we
loop over all graphs and update their visibility to match the current configs.

This updates that loop to prevent our customDrawCallback from firing when we
update the visibility. The customDrawCallback handles syncing the zoom of all
graphs and updating the start/end dates in the URL (and a few other things.)
However when a config is toggled it doesn't change any of these properties so
we're wasting a lot of time in this routine doing nothing.

This makes multi-configs much faster, though toggling a config can be slow in
some cases for reasons that are currently beyond our control.

The two reasons I have found are the Dygraph functions:
 - Dygraph.Plugins.Legend.calculateEmWidthInDiv
 - Dygraph.toRGB_

Both of these functions do dom manipulation (add a div, check some property,
and remove the div.) They do this for every series, which ends up being really
expensive. This has always been slow, but becomes really noticeable in cases
where we tripled or quadroupled the number of series with mulit-configs.

The toRGB_ function takes an rgb, hsv, or any valid css color value, and
returns it as an rgb tuple. It creates a div with the specified color and then
asks the div for the rgb color so that dygraphs doesn't have to implement hsv,
hsl, etc to rgb but doing the dom manipulation to achieve this is slow. The
toRGB_ function could be memoized, or if the value passed in is rgb it could
take a fast path that doesn't require dom manipulation.

The calculateEmWidthInDiv function is used in the legend to draw dashed lines.
I can't figure it out why it is called so often, but a comment in dygraphs says
it needs to be.

Improving either of these things will require submitting PRs to Dygraphs. (so
probably not something I'll get to until after 1.11 release.)