Back out storing this by ref in iterator records.
The special case in rebuildIterator for a captured _this being copied into a local
variable was reinstated -- except if the iterator is a "ref" method.

Methods -- even on record objects -- are effectively "ref" methods,  Otherwise, non-ref
methods on records would be required to copy the record before doing anything with it,
which is at least inefficient, and also probably not what the user expects.

The failing test case was hpl.chpl, which was causing a range allocated on the stack to
be referenced in the iterator record.

With this change, there are 18 regressions in the standard and no-local configurations,
and 6 in multilocale.  All of the standard regressions are memory leaks.