Bug fix for multi-conf graphs when used with graph expansion
Currently only affected compiler performance graphs.

If the 'default' configuration (configuration listed first) wasn't selected
when graphs were expanded, the colors on the expanded graph were all black.

This is because we grabbed the color the default graph and made all the
expanded series that color, but grabbing the color from an invisible series
returns null.