Update annotations to work with multi-configs
We attach annotations to the x-axis, but dygraphs requires that all annotations
are attached to a specific series. We randomly pick a series to attach to in
genGraphs.

For multi-configs we picked a random series that was in the default config.
This updates our javascript so that when we go to set annotations or we changed
which configurations are available, the series that annotation is attached to
is updated.

This adds a helper routine to update which series the annotations are attached
to based on the currently displayed configurations. The helper only updates the
annotations data, it does _not_ tell dygraphs to redraw with the new data. The
function has early returns to prevent doing work in cases where there's no
annotations, no series to attach to, or the series the annotations are attached
to is already displayed.

When we change which configs are visible is checks if annotations were set in
the first place, and if they were it calls the helper routine.