Update chpl_launchcmd.py to wait a couple seconds for output file before raising.
We have observed sporadic missing output files when running with PBS
launcher. @ronawho and I think we might be be hitting a slight period of time
where qstat reports the job complete, but the output file does not yet show up
in the file system (due to network latency, etc).

Update chpl_launchcmd.py to check for the output file for a couple seconds
after the job completes.