Update rangeAccessors test.
This (future) test is a feature request for write-access to individual fields in a range
object.

The original code was a sketch -- it was not expected to compile.  However, syntax errors
and a repeated variable name in the original obscured the intent of the test.

Since the feature request was entered, the question of whether strings should be mutable
has come up.  The same arguments apply to ranges:  Keeping the fields of a record
immutable may enable an optimization, i.e. the freedom to make locally-cached copies
without tracking the state of the original.

The optimization is only available if ranges are immutable at the language level.
Therefore, the decision on whether ranges are mutable or not and the rationale behind this
should be stated clearly in the Specification.  A quick scan of the 0.96 version built on
2015-02-23 apparently lacked such language.

If the decision is that ranges are immutable and the spec updated appropriately, then this
test/future can be removed from the test base.