Fixing a default sparse domain idxType bug and filing a test against it
Kyle mentioned yesterday that the default sparse domain map
doesn't do well when the idxType is something other than  'int'
(or perhaps, something that doesn't coerce to 'int', like,
specifically 'uint').  The fix turned out to be trivial,
and I filed a test against the issue so that it doesn't crop
up again in the future.