Fix removeWrapRecords, so it gets rid of all uses of the _valueType argument.
The existing version of removeWrapRecords leaves shrapnel containing references to a
removed _valueType argument.  This patch removes those statements before removing the
argument itself, so there are no remaining references to the removed _valueType argument.

This problem only surfaced with inlining turned off, since mostly the other references
were used for setting up the arguments to an inner constructor call.  With inlining turned
on, the fields were set directly, so the inner constructor call and field definitions were
not present.

This patch also allows removeWrapRecords to be run when inlining is turned off.