Compiler changes to improve performance of programs with communication
Previously, references to classes were almost always replaced with references
to wide classes. This patch allows both versions to be used based on compiler
analysis.

A static callgraph is created and used to determine if module-scope variables
can be narrowed. Nodes in the graph include information about whether or not
the function is called from within an on-clause. If a module-scope variable is
never used within one of these functions, it can be narrowed.

  A new pragma is also available, "local field", which can be applied to fields
that are classes. This is a way for the class designer to assert that the field
shares the same locale as the parent aggregate type. For the foreseeable
future we only expect to use this pragma in internal modules.
  When writing to a local field, an assertion is inserted to make sure the RHS
of the assignment shares the same locale as the parent aggregate type. These
assertions can be turned off using the "--no-local-checks" flag.
  Future work involves replacing this pragma with a keyword.

The 'narrowWideReferences' pass sets up dependencies between variables to later
determine their wideness. This analysis defaulted to "wide" when encountering
an unknown case. This patch adds more cases to avoid this default behavior.
While not every new case will result in narrowing, they should make future
development of this pass easier.