Remove overrideDefaultArr.future.
I stumbled across this in AMM testing.
The test case has been working correctly for some time, but the desired error message in
the .good file did not match the one being produced by the compiler.
The actual compiler error message appears to convey the main intent shown in the desired
output.