Revise removeWrapRecords to avoid generating bogus C code.
On the AMM development branch, I was getting invalid C code after adding memory management
for unwrapped distributions.

I traced the problem back to removeWrapRecords, which was leaving in some (newly-added)
autoCopy calls that were using a _valueType formal as the actual in the call.  The clause
in removeWrapRecords starting at line 88 (old) used to remove the _valueType actual from
all call sites and then remove the _valueType formal from the called function.

In the subject function, references to _valueType can still be left in the code if
functions that *it* calls are not similarly named "_valueType".  After the _valueType
formal is removed, the actual referring to it probably codegens to random junk.

I fixed the problem by splitting the steps that remove actuals bound to _valueType formals
from the removal of the _valueType formal.  I then added code to go through and remove all
mentions of a _valueType formal are removed from the function body.

Performing the removals in this order ensures that calls (e.g. constructor calls) that
have a _valueType argument alongside another argument will remain in the code after the
_valueType argument has been removed, but that no dangling references to a removed
_valueType formal remain in any function so modified.

Since calls can be removed from the tree on line 68, I added the guard on line 127 (new)
to avoid processing dead (and possibly corrupted) CallExprs.