Make graph selection pane more robust with scrolling and browser zoom
The graph selection pane size changes when the browser is zoomed, but we never
updated the position of the graphs. The graphs should always be next to the
selection pane, but when you zoomed their position wasn't updated.

Some browsers (I saw this with chrome) allow you to "scroll" off the end of a
window. I only ran into this with a touchpad, not by scrolling with a keyboard
or mouse. When scrolling past the window it moves the entire window and shows
gray space off to the sides. However, we were still moving the graph selection
pane while the other doms were not being moved which led to weird overlap.