Add support for 'empty' suites in GRAPHFILES and spaces before suite names
'empty' suites are represented in the GRAPHFILES with '# suite: <empty>'. Note
that the split on suite now looks for ': ' instead of ':' so that we don't
shift over every suite by a space in the html. This also means that the
formatting of suites in GRAPHFILES isn't a lax as it used to be since spaces
used to be stripped. Now they are only stripped from the right side.

This is so that it makes it easy to "group" suites inside of the suite
selection dropdown.

With xc graphs right now we have the suites:

All (perf)
NAS Parallel Benchmarks (perf)
HPCC (perf)
All (time)
NAS Parallel Benchmarks (time)
HPCC (time)
DOE (time)
Studies (time)

There is a logical grouping and sub-groping of these suites, but it's really
lost without some sort of indentation or spacing. With this we make the suite
selection dropdown look like:

All (perf)
 - NAS Parallel Benchmarks (perf)
 - HPCC (perf)

All (time)
 - NAS Parallel Benchmarks (time)
 - HPCC (time)
 - DOE (time)

Studies (time)

which makes the logical groups much easier to see.