Don't allow commas in normalizeForURL()
I started allowing commas in normalizeForURL() so you could do things like:
  normalizeForURL(arr.join())

I didn't think there would ever be any strings in individual elements of the
array but some of our graphs have commas in them, so this was a bad assumption
to make. To get around this I reverted the behavior of normalizeForURL to only
allow alpha numeric values and I added a function to convert an array of
strings to a comma separated string where each element was normalized first
(which is really what I should have done in the first place.)

I updated the previous use of normalizeForURL(arr.join()) to use the new
function, and converted some clunky code that creates a comma separated string
from an array manually to use the new function too.