Add support to 'reset' the stoke pattern for mulit-config graphs
When you have 3 or more configs on a graph, it becomes almost impossible to
distinguish some of the line (stroke pattern) types. The dashed, dot dashed, and
dashed dot look pretty much the same.

This adds an experimental button that will reset the stroke patterns based on
the currently visible configs (normally we generate the stroke patterns for all
all configs, and some of them are just hidden if that config isn't being
displayed.)

This current version is a little clunky since it doesn't maintain state between
redraws or anything (you have to click the 'reset stroke pattern' button after
every redraw, or if you change configurations.) This is really a test to see if
this sort of feature is useful, it can be improved later.