Add experimental support for an invert stroke pattern and colors buttons
For multi-config graphs like multi-locale performance we only have 1 series per
graph (with multiple configs for that series.) In this case have the same line
color for each config is pretty unhelpful.

This adds an experimental feature to use different colors for each config (and
the same stroke pattern for series if you do have multiple series on a graph.)

This is only used on graphs where there are 3 or more configurations. All the
functions are clearly marked as experimental in our javascript. This does not
affect our 'regular' graphs (chap04, chap01, etc.) only xc multi-locale and xc
single locale