Update docs() pass to use a "working temp dir" for intermediate files.
Update docs() pass to generate sphinx project and rst files in a temp dir. The
html output is then places in the root of the regular docs/ output location (or
whatever was specified in --docs-dir). This means docs/index.html just needs to
be opened (instead of docs/html/build/html/index.html).