Skip tests with stdin redirection in execopts when --nostdinredirect is set
We already skip tests with .stdin or STDIN files when --nostdinredirect is set
(either explicitly thrown or inferred from launchers we know won't work.)

I noticed that regex dna wasn't running for single locale performance testing
even through the slurm launcher supports stdin redirection. I found out that
the tests had skipifs for launcher != none to avoid running on launchers since
most of them didn't support stdin redirection.

A better approach might have been to skipif based on CHPL_NO_STDIN_REDIRECT.
Rather than add skipifs to all tests with '<' in their execopts (theres ~10
currently) I decided to update start_test to do it automatically. We already
had a section that pulled the redirect out of the execopts list so it was a
simple matter of detecting when CHPL_NO_STDIN_REDIRECT is true.

The place where this happened is after the test is compiled. It is a little
unfortunate that we compile the test only to skip it. In order to prevent
this, the logic for combining all the places execpopts come from and checking
for '<' would have to be factored out or duplicated

This also removes the CHPL_LAUNCHER != none" skipif for regexdna