Remove chpl_gasnet_configs.py.
The work for the third-party/gasnet package will be more complicated
than it was for the others, and I think I should get this in before the
weekend, so for now I'm removing it.  (The only change was a start on
the chpl_gasnet_configs.py script.  None of the supporting changes in
other places had been done.)