Improve messaging around chpldoc html output files.
Do not mention index.html. The user can figure it out. Or, if users do bump
into this, we can add a better message.

When the user specifies a --docs-dir, use it explicitly in the message (as
before). If --docs-dir is not specified, default to getCwd() + "/docs" insteads
of simply "docs". This should not change where the directory actually ends up,
but it makes the messaging nicer. E.g.

    HTML files are at: /Users/tvandoren/src/chapel/docs

Instead of:

    HTML files are at: docs

This is especially useful when calling `make module-doc`, since the "docs"
there is actually under modules/. With the new messaging, it will direct the
user to /path/to/chpl/home/modules/docs, which much more useful.