Tweak some arrayInClass tests.
On Wednesday, the test overrideDefaultArr was un-futurized.  But it depends on an array
message being generated from bounds checking.  When --fast is tossed, bounds-checking is
turned off.  In that case, this test will not produce the expected error message.

Added a .skipif file so we now avoid running the test when --fast is tossed.

The output for genericArrayInClass-otharrs-inline-iterators changed -- most likely due to
#1381 (Standalone parallel iterators for sparse and assoc).  The observed change in the
output is:

-obj/ChapelArray.c: In function 'coforall_fn9':
+obj/ChapelArray.c: In function 'coforall_fn7':

It seems plausible that the addition of standalone parallel iterators could have changed
the number or ordering of coforall functions in the generated code, and that would have
led to a renumbering and the change in the output.

I just updated the .bad file to track the observed output.  This test is run in the numa
configuration only -- which explains why the .bad file mismatch was only observed for that
configuration.  That fact also means that this change will not cause .bad file mismatches
to show up in other configurations.  (It's effectively a very surgical fix.)