Handle remote references explicitly when reading
The read function used to be inout, but I changed that with
pull request #1362. That caused the test
 multilocale/deitz/needMultiLocales/test_remote_file_read
to fail because the local portion of a remote wide pointer
would be passed to a C function. I had previously fixed that
in r21753, but had since forgotten. Because changing these
to use 'ref' intent also solves other problems, I added code
to explicitly handle the remote-ref case by making a local
temporary inside the on block.

Tested io/ferguson and test_remote_file_read with GASNET
Tested io/ferguson with COMM=none