Move the check for src==dst from comm_get to array op=
This should hopefully fix minor (10-15%) performance loss
from #746. The problem was calls to memcpy with src==dst,
which valgrind reported as overlapping memcpy. #746 and
subsequent commits fixed that by introducing chpl_memcpy
to include assert(src!=dst) and by adding a check to
chpl_gen_comm_get to do nothing in the case that src==dst.

This patch removes the check in chpl_gen_comm_get, since
the assert should do nothing in performance runs (since
the assert macro should be defined to do nothing and
since chpl_memcpy is inlined). Instead, it puts the
check in the op= overload in ChapelArray. It might one
day make sense to put an optional warning there.

I observed about a 10% speedup in promotedop test with --no-local
with this patch.