Double compilation timeout for testing
One of our biggest sources of testing noise right now is compilation timeouts.
These timeouts don't indicate a real issue, they are on systems where we run
heavily oversubscribed and on login nodes where other users could be pegging
the node we're compiling on.

Currently we only have one timeout value for compilation and timeout. Ideally
we want to split these out in sub_test but that change and threading it through
will take more time. This patch just doubles whatever the timeout value is for
compiling. So by default, comp timeout will be 10 minutes.

This is really meant a quick and dirty fix to quiet a big source of testing
noise and I'm planning on adding the more elegant solution next week (and will
try really hard to get to this.)